python-ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
Pillow==10.1.0
numpy==1.26.2
python-multipart==0.0.6
psycopg2-binary==2.9.9
opencv-python-headless==4.8.1.78
face-recognition==1.3.0
ollama==0.1.6

